+++ SpaceGamePrelim/Objects/Drawable.h
@@ -50,7 +50,7 @@ public:
 	// all inheriting classes need to work with these variables
 	void SetVisible(bool Visible) { m_Visible = Visible; }
 	Rect GetDimensions() { return m_Dimensions; }
-	Rect SetDimensions(Rect Dim) { m_Dimensions = Dim; m_LocatableScreenPosition = Vector(Dim.TopLeftX(), Dim.TopLeftY()); }
+	Rect SetDimensions(Rect Dim) { m_Dimensions = Dim; m_LocatableScreenPosition = Vector((float)Dim.TopLeftX(), (float)Dim.TopLeftY()); }
 	void SetParentScene(Scene* ParentScene);
 	Scene* GetParentScene() { return m_ParentScene; }
